+++ web/components/add-liquidity-panel.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react'
 import { Contract } from 'common/contract'
 import { formatMoney } from 'common/util/format'
 import { useUser } from 'web/hooks/use-user'
-import { addLiquidity } from 'web/lib/firebase/api-call'
+import { addLiquidity } from 'web/lib/firebase/fn-call'
 import { AmountInput } from './amount-input'
 import { Row } from './layout/row'
