+++ web/components/folds/create-fold-button.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router'
 import { useState } from 'react'
 import { PlusCircleIcon } from '@heroicons/react/solid'
 import { parseWordsAsTags } from 'common/util/parse'
-import { createFold } from 'web/lib/firebase/api-call'
+import { createFold } from 'web/lib/firebase/fn-call'
 import { foldPath } from 'web/lib/firebase/folds'
 import { toCamelCase } from 'common/util/format'
 import { ConfirmationButton } from '../confirmation-button'
