+++ web/components/numeric-resolution-panel.tsx
@@ -6,7 +6,7 @@ import { User } from 'web/lib/firebase/users'
 import { NumberCancelSelector } from './yes-no-selector'
 import { Spacer } from './layout/spacer'
 import { ResolveConfirmationButton } from './confirmation-button'
-import { resolveMarket } from 'web/lib/firebase/api-call'
+import { resolveMarket } from 'web/lib/firebase/fn-call'
 import { NumericContract } from 'common/contract'
 import { BucketInput } from './bucket-input'
